backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
python-dotenv==1.0.1
loguru==0.7.2
SQLAlchemy==2.0.35
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# === Phase 1: Image Detection ===
# Install torch separately with CPU index first (see README): pip install torch==2.4.1 --index-url https://download.pytorch.org/whl/cpu
torch==2.4.1
torchvision==0.19.1
transformers==4.44.2
Pillow>=10.4.0
numpy>=1.26,<3
opencv-python==4.10.0.84
grad-cam==1.5.4
